[TensorExpr] Make 'Tensor' a value type.
This is another commit in transition from KernelArena memory management.
Tensor is essentially just a pair of <BufPtr, StmtPtr> and we don't need
to dynamically allocate it at all - it's cheap to pass it by value, and
that's what we're switching to in this commit.

After this change nothing uses KernelScope/KernelArena and they can be
safely removed.

ghstack-source-id: 8126be6d71d89bd715ffd19876433cb9f242b724
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63586